Add some more info & docstrings

+++ run.py
@@ -14,7 +14,10 @@ parameter_file_list = []
 command_list = []
 
 def exe(command):
-    subprocess.call(command, shell=True)
+    if args.dryRun:
+        print(command)
+    else:
+        subprocess.call(command, shell=True)
 
 def mkdir():
     command = 'mkdir -p output/config output/result'
@@ -27,8 +30,8 @@ def generate_parameters():
     '''
     rA, rB, rC = myRanges # pass elements through tuple
     for i in rA:
-        for j in rC:
-            for k in rB:
+        for j in rB:
+            for k in rC:
                 str_values = "%.1f, %.1f, %.1f" % (i, j, k)
                 filename = "./output/config/input_namelist_%d_%d_%d.txt" % (i, j, k)
                 parameters = cfg.template.format(VALUES=str_values) # string format method
@@ -36,12 +39,17 @@ def generate_parameters():
                 #print(filename+'\n')
                 #print(parameters)
 
-                with open(filename, 'w') as fout:
-                    fout.write(parameters+'\n')
+                if args.dryRun:
+                    print(filename + " will be created")
+                else:
+                    with open(filename, 'w') as fout:
+                        fout.write(parameters+'\n')
 
                 parameter_file_list.append(filename)
 
-        break
+        break # only len(rB)xlen(rC) of parameter files will be created
+
+    ''' Caveat: one needs to avoid more than 1,000 files in a directory '''
 
 def create_commands():
     ''' Create commands from a list of text files with generated parameters '''
@@ -50,15 +58,23 @@ def create_commands():
         command_list.append(command)
 
 def submit_jobs():
-    ''' Submit jobs through python multiprocessing module if not in dryRun; otherwise simply pring commands '''
+    ''' Submit jobs through python multiprocessing module if not in dryRun mode; otherwise, simply printing commands '''
     if args.dryRun:
         for command in command_list: print(command)
     else:
-        pu.submit_jobs(command_list, 10)
+        pu.submit_jobs(command_list, 10) # processing 10 jobs in the same time
 
+        '''Remark: it will be even better to use bsub or condor jobs for multi-processing'''
+
+def print_final_message():
+    if args.dryRun:
+        print("[INFO] this is the end of dryRun mode (print commands only)")
+    else:
+        print("[INFO] all jobs completed!")
 
 if __name__ == "__main__":
     mkdir()
     generate_parameters()
     create_commands()
     submit_jobs()
+    print_final_message()
